requirements.txt
customtkinter==5.2.1
sqlalchemy==2.0.23
pymysql==1.1.0
bcrypt==4.1.2
cryptography==41.0.7
pillow==10.1.0
